+++ include/JoSIM/Output.hpp
@@ -13,7 +13,7 @@ namespace JoSIM {
     public:
     std::string name_;
     char type_;
-    int fileIndex = -1;
+    int64_t fileIndex = -1;
     std::vector<double> data_;
     Trace(const std::string& name) {
       name_ = name;
@@ -32,10 +32,10 @@ namespace JoSIM {
     void format_csv_or_dat(const std::string& filename,
       const char& delimiter,
       bool argmin = true,
-      int fIndex = -1);
+      int64_t fIndex = -1);
 
     void format_raw(const std::string& filename, bool argmin = true, 
-      int fIndex = -1);
+      int64_t fIndex = -1);
 
     void format_cout(const bool& argMin);
   };
